Fixed compilation warning on amd (x86_64)
git-svn-id: svn://svn.code.sf.net/p/statifier/code/statifier/trunk@73 13c1b854-16e0-4214-857e-bfa83f13b894

+++ src/phdrs.c
@@ -92,8 +92,8 @@ void *my_malloc(size_t size, const char *item, const char *pgm_name)
 	if (result == NULL) {
 		fprintf(
 			stderr,
-			"%s: Can't malloc %d byte for '%s'.\n",
-			pgm_name, size, item
+			"%s: Can't malloc %lu byte for '%s'.\n",
+			pgm_name, (unsigned long)size, item
 		);
 		return NULL;
 	}
